refactor(problem7): 내부 클래스들을 별도의 외부 클래스로 분리
